consolidate gated mlp
