+++ models/tt_dit/models/transformers/diffusion_gemma/self_conditioning.py
@@ -3,25 +3,22 @@
 
 import ttnn
 
-from ....layers.linear import ColParallelLinear, RowParallelLinear
+from ....layers.feedforward import GatedMLP
 from ....layers.module import Module
 from ....layers.normalization import RMSNorm
 from ....parallel.config import DiTParallelConfig
 
-TILE = ttnn.TILE_SIZE
-
 
 class DiffusionGemmaSelfConditioning(Module):
-    """Gated-MLP self-conditioning block prepended to the decoder.
+    """Self-conditioning block prepended to the decoder.
 
-    Reference: transformers.models.diffusion_gemma.modeling_diffusion_gemma.DiffusionGemmaSelfConditioning.
+    Reference: ``transformers.models.diffusion_gemma.modeling_diffusion_gemma.DiffusionGemmaSelfConditioning``::
 
         normed = pre_norm(self_conditioning_signal)
         sc     = down_proj( act_fn(gate_proj(normed)) * up_proj(normed) )
         return post_norm(inputs_embeds + sc)
 
-    The input embeddings are passed through replicated; the gated MLP is megatron-style
-    TP (column-parallel on the gate/up, row-parallel on the down). Output is replicated.
+    The gated body is the shared ``GatedMLP`` primitive.
     """
 
     def __init__(
@@ -35,24 +32,9 @@ class DiffusionGemmaSelfConditioning(Module):
         parallel_config: DiTParallelConfig,
     ) -> None:
         super().__init__()
-
-        self.hidden_size = hidden_size
-        self.intermediate_size = intermediate_size
-
-        tp_factor = parallel_config.tensor_parallel.factor
-        # Tile alignment: per-device intermediate dim must be a multiple of 32.
-        # DiffusionGemma uses intermediate_size=2112 → divisible by 32 only for tp∈{1,2}.
-        # Larger TP factors will require padding handled by a higher-level wrapper.
-        assert (intermediate_size // tp_factor) % TILE == 0, (
-            f"intermediate_size ({intermediate_size}) / tp_factor ({tp_factor}) "
-            f"must be tile-aligned ({TILE}); add output padding if larger TP is needed."
-        )
-        assert hidden_size % TILE == 0
-
         self.parallel_config = parallel_config
         self.mesh_device = mesh_device
 
-        # pre_norm: weight, no bias. Input is replicated, normed over hidden_size → use plain RMSNorm.
         self.pre_norm = RMSNorm(
             embedding_dim=hidden_size,
             norm_eps=rms_norm_eps,
@@ -60,7 +42,13 @@ class DiffusionGemmaSelfConditioning(Module):
             bias=False,
             mesh_device=mesh_device,
         )
-        # post_norm: no scale (with_scale=False in HF) → norm_elementwise_affine=False.
+        self.gated_mlp = GatedMLP(
+            hidden_size=hidden_size,
+            intermediate_size=intermediate_size,
+            mesh_device=mesh_device,
+            ccl_manager=ccl_manager,
+            parallel_config=parallel_config,
+        )
         self.post_norm = RMSNorm(
             embedding_dim=hidden_size,
             norm_eps=rms_norm_eps,
@@ -69,44 +57,20 @@ class DiffusionGemmaSelfConditioning(Module):
             mesh_device=mesh_device,
         )
 
-        col_kwargs = dict(
-            bias=False,
-            mesh_device=mesh_device,
-            mesh_axis=parallel_config.tensor_parallel.mesh_axis,
-            ccl_manager=ccl_manager,
-        )
-        self.gate_proj = ColParallelLinear(hidden_size, intermediate_size, activation_fn="gelu_tanh", **col_kwargs)
-        self.up_proj = ColParallelLinear(hidden_size, intermediate_size, **col_kwargs)
-        self.down_proj = RowParallelLinear(
-            intermediate_size,
-            hidden_size,
-            bias=False,
-            mesh_device=mesh_device,
-            mesh_axis=parallel_config.tensor_parallel.mesh_axis,
-            ccl_manager=ccl_manager,
-        )
+    def _prepare_torch_state(self, state: dict) -> None:
+        """HF stores ``gate_proj``/``up_proj``/``down_proj`` as direct children of
+        ``DiffusionGemmaSelfConditioning``; we nest them under ``gated_mlp``. Re-prefix."""
+        for name in ("gate_proj", "up_proj", "down_proj"):
+            for k in list(state.keys()):
+                if k == f"{name}.weight" or k.startswith(f"{name}."):
+                    state[f"gated_mlp.{k}"] = state.pop(k)
 
-    def forward(
-        self,
-        inputs_embeds: ttnn.Tensor,
-        self_conditioning_signal: ttnn.Tensor,
-        *,
-        compute_kernel_config=None,
-    ) -> ttnn.Tensor:
-        """
-        inputs_embeds, self_conditioning_signal: replicated [B, canvas_length, hidden_size].
-        Output: replicated [B, canvas_length, hidden_size].
-        """
-        normed = self.pre_norm(self_conditioning_signal, compute_kernel_config=compute_kernel_config)
-        gate = self.gate_proj(normed, parallel_config=self.parallel_config, compute_kernel_config=compute_kernel_config)
-        up = self.up_proj(normed, parallel_config=self.parallel_config, compute_kernel_config=compute_kernel_config)
-        gated = ttnn.multiply(gate, up)
-        ttnn.deallocate(gate)
-        ttnn.deallocate(up)
-        sc_signal = self.down_proj(gated, compute_kernel_config=compute_kernel_config)
-        ttnn.deallocate(gated)
+    def forward(self, inputs_embeds: ttnn.Tensor, self_conditioning_signal: ttnn.Tensor) -> ttnn.Tensor:
+        normed = self.pre_norm(self_conditioning_signal)
+        sc_signal = self.gated_mlp(normed)
+        ttnn.deallocate(normed)
         combined = ttnn.add(inputs_embeds, sc_signal)
         ttnn.deallocate(sc_signal)
-        out = self.post_norm(combined, compute_kernel_config=compute_kernel_config)
+        out = self.post_norm(combined)
         ttnn.deallocate(combined)
         return out
